requirements.txt
numpy 
pandas
requests
python-dotenv
matplotlib
pyarrow
fastparquet
statsmodels
scikit-learn
sqlalchemy
psycopg2-binary
supabase